requirements.txt
disnake
aiohttp
colorama
python-dotenv
psutil
humanize
motor[srv]